+++ instrumentation/opentelemetry-instrumentation-botocore/tests/conftest.py
@@ -3,12 +3,12 @@
 import json
 import os
 
+import boto3
 import pytest
 import yaml
-from openai import AsyncOpenAI, OpenAI
 
-from opentelemetry.instrumentation.openai_v2 import OpenAIInstrumentor
-from opentelemetry.instrumentation.openai_v2.utils import (
+from opentelemetry.instrumentation.botocore import BotocoreInstrumentor
+from opentelemetry.instrumentation.botocore.environment_variables import (
     OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT,
 )
 from opentelemetry.sdk._events import EventLoggerProvider
@@ -16,17 +16,6 @@ from opentelemetry.sdk._logs import LoggerProvider
 from opentelemetry.sdk._logs.export import (
     InMemoryLogExporter,
     SimpleLogRecordProcessor,
-)
-from opentelemetry.sdk.metrics import (
-    Histogram,
-    MeterProvider,
-)
-from opentelemetry.sdk.metrics.export import (
-    InMemoryMetricReader,
-)
-from opentelemetry.sdk.metrics.view import (
-    ExplicitBucketHistogramAggregation,
-    View,
 )
 from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import SimpleSpanProcessor
@@ -47,12 +36,6 @@ def fixture_log_exporter():
     yield exporter
 
 
-@pytest.fixture(scope="function", name="metric_reader")
-def fixture_metric_reader():
-    exporter = InMemoryMetricReader()
-    yield exporter
-
-
 @pytest.fixture(scope="function", name="tracer_provider")
 def fixture_tracer_provider(span_exporter):
     provider = TracerProvider()
@@ -69,76 +52,21 @@ def fixture_event_logger_provider(log_exporter):
     return event_logger_provider
 
 
-@pytest.fixture(scope="function", name="meter_provider")
-def fixture_meter_provider(metric_reader):
-    token_usage_histogram_view = View(
-        instrument_type=Histogram,
-        instrument_name="gen_ai.client.token.usage",
-        aggregation=ExplicitBucketHistogramAggregation(
-            boundaries=[
-                1,
-                4,
-                16,
-                64,
-                256,
-                1024,
-                4096,
-                16384,
-                65536,
-                262144,
-                1048576,
-                4194304,
-                16777216,
-                67108864,
-            ]
-        ),
-    )
-
-    duration_histogram_view = View(
-        instrument_type=Histogram,
-        instrument_name="gen_ai.client.operation.duration",
-        aggregation=ExplicitBucketHistogramAggregation(
-            boundaries=[
-                0.01,
-                0.02,
-                0.04,
-                0.08,
-                0.16,
-                0.32,
-                0.64,
-                1.28,
-                2.56,
-                5.12,
-                10.24,
-                20.48,
-                40.96,
-                81.92,
-            ]
-        ),
-    )
-
-    meter_provider = MeterProvider(
-        metric_readers=[metric_reader],
-        views=[token_usage_histogram_view, duration_histogram_view],
-    )
-
-    return meter_provider
+@pytest.fixture
+def bedrock_runtime_client():
+    return boto3.client("bedrock-runtime")
 
 
 @pytest.fixture(autouse=True)
 def environment():
-    if not os.getenv("OPENAI_API_KEY"):
-        os.environ["OPENAI_API_KEY"] = "test_openai_api_key"
-
-
-@pytest.fixture
-def openai_client():
-    return OpenAI()
-
-
-@pytest.fixture
-def async_openai_client():
-    return AsyncOpenAI()
+    if not os.getenv("AWS_ACCESS_KEY_ID"):
+        os.environ["AWS_ACCESS_KEY_ID"] = "test_aws_access_key_id"
+    if not os.getenv("AWS_SECRET_ACCESS_KEY"):
+        os.environ["AWS_SECRET_ACCESS_KEY"] = "test_aws_secret_key"
+    if not os.getenv("AWS_SESSION_TOKEN"):
+        os.environ["AWS_SESSION_TOKEN"] = "test_aws_session_token"
+    if not os.getenv("AWS_DEFAULT_REGION"):
+        os.environ["AWS_DEFAULT_REGION"] = "eu-central-1"
 
 
 @pytest.fixture(scope="module")
@@ -146,9 +74,8 @@ def vcr_config():
     return {
         "filter_headers": [
             ("cookie", "test_cookie"),
-            ("authorization", "Bearer test_openai_api_key"),
-            ("openai-organization", "test_openai_org_id"),
-            ("openai-project", "test_openai_project_id"),
+            ("authorization", "Bearer test_aws_authorization"),
+            ("X-Amz-Security-Token", "test_aws_security_token"),
         ],
         "decode_compressed_response": True,
         "before_record_response": scrub_response_headers,
@@ -156,18 +83,15 @@ def vcr_config():
 
 
 @pytest.fixture(scope="function")
-def instrument_no_content(
-    tracer_provider, event_logger_provider, meter_provider
-):
+def instrument_no_content(tracer_provider, event_logger_provider):
     os.environ.update(
         {OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT: "False"}
     )
 
-    instrumentor = OpenAIInstrumentor()
+    instrumentor = BotocoreInstrumentor()
     instrumentor.instrument(
         tracer_provider=tracer_provider,
         event_logger_provider=event_logger_provider,
-        meter_provider=meter_provider,
     )
 
     yield instrumentor
@@ -176,17 +100,14 @@ def instrument_no_content(
 
 
 @pytest.fixture(scope="function")
-def instrument_with_content(
-    tracer_provider, event_logger_provider, meter_provider
-):
+def instrument_with_content(tracer_provider, event_logger_provider):
     os.environ.update(
         {OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT: "True"}
     )
-    instrumentor = OpenAIInstrumentor()
+    instrumentor = BotocoreInstrumentor()
     instrumentor.instrument(
         tracer_provider=tracer_provider,
         event_logger_provider=event_logger_provider,
-        meter_provider=meter_provider,
     )
 
     yield instrumentor
@@ -265,6 +186,5 @@ def scrub_response_headers(response):
     """
     This scrubs sensitive response headers. Note they are case-sensitive!
     """
-    response["headers"]["openai-organization"] = "test_openai_org_id"
     response["headers"]["Set-Cookie"] = "test_set_cookie"
     return response